Pipeline04/requirements.txt
# Manipulação de dados
pandas==2.2.2

# Escrita e leitura de arquivos Parquet
pyarrow==15.0.2

# Integração com AWS S3
boto3==1.34.57
s3fs==2024.6.0
fsspec==2024.6.0

# Variáveis de ambiente e configuração
python-dotenv==1.0.1

# Logging estruturado
loguru==0.7.2

#yaml
pyyaml